move common utils to particula dir
