+++ particula/utils/_common/tests/constants_test.py
@@ -2,7 +2,7 @@
 """
 
 from particula import u
-from particula.utils._common.get_constants import (
+from particula.get_constants import (
     BOLTZMANN_CONSTANT,
     AVOGADRO_NUMBER,
     GAS_CONSTANT,
